+++ packages/contracts/src/sessions/procedures/index.ts
@@ -2,3 +2,4 @@ export * from "./create-session";
 export * from "./get-session";
 export * from "./list-composer-suggestions";
 export * from "./list-models";
+export * from "./rename-session";